refactor: move config to common section
