+++ src/api/app.js
@@ -29,7 +29,7 @@ import {get as _get} from 'lodash';
 import appCleanup from '../server/helpers/appCleanup';
 import bodyParser from 'body-parser';
 import compression from 'compression';
-import config from '../server/helpers/config';
+import config from '../common/helpers/config';
 import express from 'express';
 import favicon from 'serve-favicon';
 import git from 'git-rev';
