+++ src/server/helpers/auth.js
@@ -22,7 +22,7 @@ import * as error from '../helpers/error';
 
 import Log from 'log';
 import _ from 'lodash';
-import config from './config';
+import config from '../../common/helpers/config';
 import passport from 'passport';
 import status from 'http-status';
